# Genericize SIMD arch code with `simd8<T>`