--- a/amalgamation.sh
+++ b/amalgamation.sh
@@ -20,12 +20,9 @@ $SCRIPTPATH/src/simdjson.cpp
 $SCRIPTPATH/src/jsonioutil.cpp
 $SCRIPTPATH/src/jsonminifier.cpp
 $SCRIPTPATH/src/jsonparser.cpp
-$SCRIPTPATH/src/arm64/simd_input.h
-$SCRIPTPATH/src/haswell/simd_input.h
-$SCRIPTPATH/src/westmere/simd_input.h
-$SCRIPTPATH/src/arm64/simdutf8check.h
-$SCRIPTPATH/src/haswell/simdutf8check.h
-$SCRIPTPATH/src/westmere/simdutf8check.h
+$SCRIPTPATH/src/arm64/simd.h
+$SCRIPTPATH/src/haswell/simd.h
+$SCRIPTPATH/src/westmere/simd.h
 $SCRIPTPATH/src/arm64/stage1_find_marks.h
 $SCRIPTPATH/src/haswell/stage1_find_marks.h
 $SCRIPTPATH/src/westmere/stage1_find_marks.h