This replaces all architecture-specific simdutf8check.h's with a common generic simdutf8check.h that relies on new `simd8` architecture-specific primitives (it also genericizes the other arch-specific code with the same primitives, but there is less code there already :)).

I had these goals:
* *Make cross-architecture algorithm changes easier:* hide architecture specific functions for familiar operations like &, | and lookup.
* *Make new architectures easier to create:* more common code.
* *No effect on performance:* generate the exact same intrinsics.
* *Retain architecture-specific algorithm ability:* make sure we can still write architecture-specific algorithms the way we currently do, and keep the new abstractions usable with intrinsics.

Haswell perf is the same. Westmere is very slightly better (I think because of the 32-bit mask thing), and so is ARM based on CI runs (probably vqmaxq, but I can and will test that so we at least know why).

The number of LOC is about the same overall, but the surface area / conceptual load to understand to grok the algorithm is vastly reduced and these abstractions will also likely help genericize some stage 1 and stage 2 code.

## Guide

1. [`simd8<int8_t>, simd8<uint8_t>, simd8<bool>`](https://github.com/lemire/simdjson/blob/f1ab469eb7709790d756e4e18528b41e108f55a5/src/haswell/simd.h): I created an `arch/simd.h` with signed, unsigned and bool mask types similar to [rust's simd types](https://rust-lang-nursery.github.io/packed_simd/packed_simd/#vector-types). They provide bitwise and arithmetic primitives used everywhere else.

   One essential feature of the generic simd types is that they convert effortlessly into and out of the corresponding architecture-specific type (e.g. `__m256i`), so that when we do dip into architecture-specific algorithms, we can use them seamlessly with intrinsics.

   Another feature I found super useful was an implicit `splat` constructor: we do a lot of operations like `_mm_sub_epi8(prev_carried_continuations, _mm256_setr_epi8(1))`, and this lets us write them as `prev_carried_continuations - 1` rather than `prev_carried_continuations - simd::u8::splat(1)`.

  The type safety also helps: while you can implicitly convert in and out of `__mm256`, you have to explicitly cast from `simd8<int8_t>` to `simd8<uint8_t>` and vice versa so that you don't accidentally use the wrong arithmetic.

   I also renamed `simd_input` to `simd::u8x64` to make it a little clearer what it is, and removed `simd_input.h`.

2. [`generic/simdutf8check.h`](https://github.com/lemire/simdjson/blob/f1ab469eb7709790d756e4e18528b41e108f55a5/src/generic/simdutf8check.h): I genericized the utf8 check algorithm by replacing all architecture intrinsics with corresponding primitives. I did this by carefully replacing each intrinsic call in the Haswell and ARM code. For Westmere I was a tiny bit more cavalier, comparing about a third of it by hand and relying on performance and feature testing to shake out any other issues.

3. I genericized the *code* in the remaining arch files for readability and concept load, but did not move all of them into `generic` because ARM and Intel have fundamentally different algorithms for two of them.

4. I renamed compute_quote_mask to prefix_xor (because it has the effect of a bitwise cumulative xor) and moved it to an arch-specific [`bitmask.h`](https://github.com/lemire/simdjson/pull/346#issuecomment-546728931). This is the one I'm least sure about because I don't know what the term "prefix xor" would mean to someone steeped in parallel algorithms.

## Differences

I hand-checked each bit of code I replaced, and it's all the same except two things:

* westmere goes a lot faster if its `to_bitmask()` returns `uint32_t` instead of `uint16_t`, for some reason. I'm not sure if this is a fluke or what.
* arm: I changed the [series of 4 instructions](https://github.com/lemire/simdjson/blob/master/src/arm64/simdutf8check.h#L177-L180) we were using to test if an `arm64` vector was all zeroes, to `vmaxvq_u8(value) == 0`. I found it on stack overflow when trying to understand how to do the operation, and I am curious why we used the longer series before.
* In `check_carried_continuations`, we [initialize a vector with all 9's and a 1 at the end](https://github.com/lemire/simdjson/blob/master/src/haswell/simdutf8check.h#L108-L110). There isn't a single well-known primitive to perform this operation across multiple vector sizes, so I [created a const array](https://github.com/lemire/simdjson/blob/f1ab469eb7709790d756e4e18528b41e108f55a5/src/generic/simdutf8check.h#L67-L72) with 31 9's and a 1, and now use `loadu` on that instead of `setr`. It did not affect performance.