@jkeiser 2019-10-24
It looks like clang is going to need a bit of work; I did some fancy template work trying to share code between m8/u8/i8, and I may need to just do a wee bit of copy/paste instead :)