@jkeiser 2019-10-24
@lemire the requested review is just a ping to take a look at the approach--this needs clang and probably arm fixes before being ready.