@lemire 2019-10-24
@jkeiser This is nice work. Obviously, we will want to run multiplatform, multicompiler benchmarks to make sure there is no regression.

We also have alternative algorithms for UTF-8 validation that use slightly fewer instructions... there are a few outstanding PRs...

https://github.com/lemire/simdjson/pull/286

Would you have a look? We would want to make sure that these other algorithms can also fit into your model.