@jkeiser 2019-10-25
@lemire I'll take a look!

Note: even if they don't fit and won't fit, it's likely not a blocker here. I don't feel like we should try to put every *possible* primitive into a generic simd thing, just common nouns and verbs with fast implementations on all architectures--which means we'll sometimes need architecture-specific code in the algorithms themselves (the hybrid model of the rest of stage 1). cmul, for example, is not common enough that a primitive is a good idea--you don't want people using it thinking it's going to be a single instruction and finding out someone implemented it with a for loop.

As far as validating, I'll get clang and avx working first, but if you'd like to run the benchmarks for westmere and haswell when you have time, I'd appreciate it. As I said, I'm seeing haswell identical and westmere faster (the former being expected because it *should* emit the same intrinsics), but that's only on Kaby Lake.