@jkeiser 2019-10-25
Looking at it, I believe everything there fits into the existing primitives. The only new thing there is a saturated add, but I already went ahead and added that when I added saturated subtract.