@jkeiser 2019-10-25
clang looks fixed, arm64 compiles--but arm is failing. I'll look back through and see what I might have translated wrong ...