@jkeiser 2019-10-26
@lemire ready for review (all tests pass). Everything looks faster. Please do run this wherever you can to get some validation on the numbers I'm seeing: most things are a wee bit faster, Haswell on my machine stays the same. ARM perf runs on CI show it going faster as well.