@jkeiser 2019-10-27
lgtm-bot made it pretty hard to read this: it was commenting on how there is a new header without header guard (which is true and expected).