@jkeiser 2019-10-27
I just added 3 commits that complete the genericization of src/<arch>/*, as much as possible. I'll happily split it into a separate or later checkin if you so desire--I added it here so that performance checking only really has to be done once.

ARM has two algorithms that can't be combined with the Intel algorithms:

- `find_whitespace_and_operators`: I'm honestly unclear why this is slower. It seems to have roughly the same number of operations, but on ARM munging the table lookups together and bitmasking from there is objectively about 1.5% faster than doing two separate table lookups. The ARM algorithm is slower on Intel, but that may be because I wrote an inefficient implementation for `any_bits_set`.
- `find_bs_and_quote_bits`: ARM mask calculation is the same cost for anything less 64 bytes, so we want to pack as many masks into 64 bytes as possible. Using the intel algorithm for arm, or the arm algorithm for westmere, slows them down.

In both cases, I switched to using the new primitives but left the algorithms alone.