@jkeiser 2019-10-27
Re-amalgamated and re-pushed. No more changes coming, this is done and dusted :)