@lemire 2019-10-28
Got it. On my todo.