@jkeiser 2019-10-29
@lemire Ugh. Hold on, something's changed with the latest merges. Something I did when messing with ARM regressed Intel. I'll look at it tonight.