@jkeiser 2019-10-29
Yeah, circle says the same as I did a day or two ago: avx perf is identical and sse is a very tiny bit better.