@jkeiser 2019-10-29
@lemire OK, it was a false alarm, you are go for review :) My system was under load I didn't know about this morning when I ran the benchmarks; it righted itself when I killed a runaway process.