@lemire 2019-11-01
@jkeiser 

I see a systematic performance regression on stage 1 in my tests for x64 processors. Both the number of mispredicted branches and the number of instructions is up.

skylake (GNU GCC 8.3)

```
jsonexamples/apache_builds.json
PR:
stage 1 instructions:     314647 cycles:      84055 (44.42 %) ins/cycles: 3.74 mis. branches:          9 (cycles/mis.branch 9128.51) cache accesses:       4044 (failure          0)
 stage 1 runs at 0.66 cycles per input byte.
master:
stage 1 instructions:     310668 cycles:      81932 (43.57 %) ins/cycles: 3.79 mis. branches:         10 (cycles/mis.branch 7563.91) cache accesses:       3611 (failure          0)
 stage 1 runs at 0.64 cycles per input byte.

jsonexamples/canada.json
PR:
stage 1 instructions:    6212659 cycles:    1738844 (22.87 %) ins/cycles: 3.57 mis. branches:       5358 (cycles/mis.branch 324.48) cache accesses:     122974 (failure       2387)
 stage 1 runs at 0.77 cycles per input byte.
master:
stage 1 instructions:    6142312 cycles:    1707109 (22.47 %) ins/cycles: 3.60 mis. branches:       5357 (cycles/mis.branch 318.64) cache accesses:     114869 (failure       2817)
 stage 1 runs at 0.76 cycles per input byte.

jsonexamples/citm_catalog.json
PR:
stage 1 instructions:    4043681 cycles:    1078121 (46.48 %) ins/cycles: 3.75 mis. branches:        201 (cycles/mis.branch 5347.82) cache accesses:      72389 (failure         43)
 stage 1 runs at 0.62 cycles per input byte.
master:
stage 1 instructions:    3990058 cycles:    1049421 (45.76 %) ins/cycles: 3.80 mis. branches:        185 (cycles/mis.branch 5645.09) cache accesses:      77483 (failure         81)
 stage 1 runs at 0.61 cycles per input byte.

jsonexamples/github_events.json
PR:
stage 1 instructions:     150578 cycles:      40908 (43.97 %) ins/cycles: 3.68 mis. branches:          4 (cycles/mis.branch 8696.62) cache accesses:        693 (failure          0)
 stage 1 runs at 0.63 cycles per input byte.
master:
stage 1 instructions:     148551 cycles:      39763 (43.59 %) ins/cycles: 3.74 mis. branches:          4 (cycles/mis.branch 9827.93) cache accesses:        537 (failure          0)
 stage 1 runs at 0.61 cycles per input byte.

jsonexamples/gsoc-2018.json
PR:
stage 1 instructions:    6171919 cycles:    1788030 (51.10 %) ins/cycles: 3.45 mis. branches:       5621 (cycles/mis.branch 318.09) cache accesses:     114832 (failure       4749)
 stage 1 runs at 0.54 cycles per input byte.
master:
stage 1 instructions:    6067924 cycles:    1736427 (50.52 %) ins/cycles: 3.49 mis. branches:       5652 (cycles/mis.branch 307.18) cache accesses:     114395 (failure       3984)
 stage 1 runs at 0.52 cycles per input byte.

jsonexamples/instruments.json
PR:
stage 1 instructions:     558117 cycles:     149667 (40.81 %) ins/cycles: 3.73 mis. branches:         61 (cycles/mis.branch 2418.59) cache accesses:       9522 (failure          0)
 stage 1 runs at 0.68 cycles per input byte.
master:
stage 1 instructions:     551230 cycles:     145664 (39.94 %) ins/cycles: 3.78 mis. branches:         39 (cycles/mis.branch 3693.31) cache accesses:       9818 (failure          0)
 stage 1 runs at 0.66 cycles per input byte.

jsonexamples/marine_ik.json
PR:
stage 1 instructions:    9320763 cycles:    2612401 (23.31 %) ins/cycles: 3.57 mis. branches:       8516 (cycles/mis.branch 306.74) cache accesses:     182736 (failure      33691)
 stage 1 runs at 0.88 cycles per input byte.
master:
stage 1 instructions:    9247348 cycles:    2587152 (22.40 %) ins/cycles: 3.57 mis. branches:       8523 (cycles/mis.branch 303.53) cache accesses:     191340 (failure      33444)
 stage 1 runs at 0.87 cycles per input byte.

jsonexamples/mesh.json
PR:
stage 1 instructions:    2183901 cycles:     586101 (21.65 %) ins/cycles: 3.73 mis. branches:        875 (cycles/mis.branch 669.48) cache accesses:      43134 (failure          1)
 stage 1 runs at 0.81 cycles per input byte.
master:
stage 1 instructions:    2170136 cycles:     580658 (20.67 %) ins/cycles: 3.74 mis. branches:        790 (cycles/mis.branch 734.59) cache accesses:      46113 (failure          0)
 stage 1 runs at 0.80 cycles per input byte.

jsonexamples/mesh.pretty.json
PR:
stage 1 instructions:    3862363 cycles:    1024327 (28.93 %) ins/cycles: 3.77 mis. branches:         34 (cycles/mis.branch 29951.11) cache accesses:      69714 (failure         55)
 stage 1 runs at 0.65 cycles per input byte.
master:
stage 1 instructions:    3813068 cycles:     999757 (27.68 %) ins/cycles: 3.81 mis. branches:         36 (cycles/mis.branch 27167.32) cache accesses:      69796 (failure        192)
 stage 1 runs at 0.63 cycles per input byte.

jsonexamples/numbers.json
PR:
stage 1 instructions:     382808 cycles:     104285 (21.66 %) ins/cycles: 3.67 mis. branches:        126 (cycles/mis.branch 823.23) cache accesses:       5874 (failure          0)
 stage 1 runs at 0.69 cycles per input byte.
master:
stage 1 instructions:     378117 cycles:     101188 (20.12 %) ins/cycles: 3.74 mis. branches:         82 (cycles/mis.branch 1222.28) cache accesses:       5677 (failure          0)
 stage 1 runs at 0.67 cycles per input byte.

jsonexamples/random.json
PR:
stage 1 instructions:    1988441 cycles:     560493 (43.55 %) ins/cycles: 3.55 mis. branches:        977 (cycles/mis.branch 573.25) cache accesses:      29855 (failure          0)
 stage 1 runs at 1.10 cycles per input byte.
master:
stage 1 instructions:    1992950 cycles:     560980 (44.19 %) ins/cycles: 3.55 mis. branches:       1003 (cycles/mis.branch 558.80) cache accesses:      29709 (failure          0)
 stage 1 runs at 1.10 cycles per input byte.

jsonexamples/twitterescaped.json
PR:
stage 1 instructions:    1324898 cycles:     379921 (22.61 %) ins/cycles: 3.49 mis. branches:       1451 (cycles/mis.branch 261.79) cache accesses:      25145 (failure          0)
 stage 1 runs at 0.68 cycles per input byte.
master:
stage 1 instructions:    1308968 cycles:     372494 (21.78 %) ins/cycles: 3.51 mis. branches:       1461 (cycles/mis.branch 254.81) cache accesses:      26857 (failure          0)
 stage 1 runs at 0.66 cycles per input byte.

jsonexamples/twitter.json
PR:
stage 1 instructions:    1722512 cycles:     490941 (49.30 %) ins/cycles: 3.51 mis. branches:       1174 (cycles/mis.branch 418.07) cache accesses:      29305 (failure          0)
 stage 1 runs at 0.78 cycles per input byte.
master:
stage 1 instructions:    1711881 cycles:     480486 (48.61 %) ins/cycles: 3.56 mis. branches:        986 (cycles/mis.branch 486.99) cache accesses:      29140 (failure          0)
 stage 1 runs at 0.76 cycles per input byte.

jsonexamples/update-center.json
PR:
stage 1 instructions:    1340474 cycles:     393882 (38.99 %) ins/cycles: 3.40 mis. branches:       2127 (cycles/mis.branch 185.10) cache accesses:      27101 (failure          0)
 stage 1 runs at 0.74 cycles per input byte.
master:
stage 1 instructions:    1324186 cycles:     382760 (38.37 %) ins/cycles: 3.46 mis. branches:       1905 (cycles/mis.branch 200.83) cache accesses:      27118 (failure          0)
 stage 1 runs at 0.72 cycles per input byte.

```



Cannon Lake (GNU GCC 8.3)

```
jsonexamples/apache_builds.json
PR:
stage 1 instructions:     314622 cycles:      83223 (44.73 %) ins/cycles: 3.78 mis. branches:         13 (cycles/mis.branch 6309.14) cache accesses:       2590 (failure          0)
 stage 1 runs at 0.65 cycles per input byte.
master:
stage 1 instructions:     309644 cycles:      81164 (44.08 %) ins/cycles: 3.82 mis. branches:         15 (cycles/mis.branch 5384.76) cache accesses:       2804 (failure          0)
 stage 1 runs at 0.64 cycles per input byte.

jsonexamples/canada.json
PR:
stage 1 instructions:    6212636 cycles:    1779418 (23.89 %) ins/cycles: 3.49 mis. branches:       5150 (cycles/mis.branch 345.47) cache accesses:      74573 (failure      20939)
 stage 1 runs at 0.79 cycles per input byte.
master:
stage 1 instructions:    6124698 cycles:    1735775 (23.48 %) ins/cycles: 3.53 mis. branches:       5287 (cycles/mis.branch 328.30) cache accesses:      74406 (failure      20727)
 stage 1 runs at 0.77 cycles per input byte.

jsonexamples/citm_catalog.json
PR:
stage 1 instructions:    4043659 cycles:    1073688 (46.72 %) ins/cycles: 3.77 mis. branches:        215 (cycles/mis.branch 4982.31) cache accesses:      47208 (failure       3821)
 stage 1 runs at 0.62 cycles per input byte.
master:
stage 1 instructions:    3976538 cycles:    1049099 (46.22 %) ins/cycles: 3.79 mis. branches:        195 (cycles/mis.branch 5355.28) cache accesses:      47194 (failure       1815)
 stage 1 runs at 0.61 cycles per input byte.

jsonexamples/github_events.json
PR:
stage 1 instructions:     150553 cycles:      41263 (44.83 %) ins/cycles: 3.65 mis. branches:          5 (cycles/mis.branch 7432.17) cache accesses:       1106 (failure          0)
 stage 1 runs at 0.63 cycles per input byte.
master:
stage 1 instructions:     148014 cycles:      39361 (43.81 %) ins/cycles: 3.76 mis. branches:          7 (cycles/mis.branch 5466.14) cache accesses:        302 (failure          0)
 stage 1 runs at 0.60 cycles per input byte.

jsonexamples/gsoc-2018.json
PR:
stage 1 instructions:    6171894 cycles:    1791889 (50.00 %) ins/cycles: 3.44 mis. branches:       5385 (cycles/mis.branch 332.71) cache accesses:      75218 (failure      18618)
 stage 1 runs at 0.54 cycles per input byte.
master:
stage 1 instructions:    6041895 cycles:    1752034 (49.82 %) ins/cycles: 3.45 mis. branches:       5591 (cycles/mis.branch 313.35) cache accesses:      75162 (failure      19000)
 stage 1 runs at 0.53 cycles per input byte.

jsonexamples/instruments.json
PR:
stage 1 instructions:     558096 cycles:     147940 (41.35 %) ins/cycles: 3.77 mis. branches:         64 (cycles/mis.branch 2288.40) cache accesses:       6343 (failure          0)
 stage 1 runs at 0.67 cycles per input byte.
master:
stage 1 instructions:     549483 cycles:     144713 (40.95 %) ins/cycles: 3.80 mis. branches:         71 (cycles/mis.branch 2013.21) cache accesses:       6634 (failure          1)
 stage 1 runs at 0.66 cycles per input byte.

jsonexamples/marine_ik.json
PR:
stage 1 instructions:    9320738 cycles:    2638033 (23.89 %) ins/cycles: 3.53 mis. branches:       8663 (cycles/mis.branch 304.51) cache accesses:     115520 (failure      63210)
 stage 1 runs at 0.88 cycles per input byte.
master:
stage 1 instructions:    9224012 cycles:    2566720 (23.21 %) ins/cycles: 3.59 mis. branches:       8471 (cycles/mis.branch 302.99) cache accesses:     115542 (failure      64791)
 stage 1 runs at 0.86 cycles per input byte.

jsonexamples/mesh.json
PR:
stage 1 instructions:    2183884 cycles:     574164 (21.76 %) ins/cycles: 3.80 mis. branches:        618 (cycles/mis.branch 928.25) cache accesses:      27809 (failure         37)
 stage 1 runs at 0.79 cycles per input byte.
master:
stage 1 instructions:    2164461 cycles:     570213 (21.58 %) ins/cycles: 3.80 mis. branches:        621 (cycles/mis.branch 918.16) cache accesses:      27802 (failure        190)
 stage 1 runs at 0.79 cycles per input byte.

jsonexamples/mesh.pretty.json
PR:
stage 1 instructions:    3862346 cycles:    1019808 (29.12 %) ins/cycles: 3.79 mis. branches:         43 (cycles/mis.branch 23552.16) cache accesses:      45487 (failure       2958)
 stage 1 runs at 0.65 cycles per input byte.
master:
stage 1 instructions:    3800723 cycles:     992963 (28.53 %) ins/cycles: 3.83 mis. branches:         39 (cycles/mis.branch 25138.33) cache accesses:      45504 (failure       1835)
 stage 1 runs at 0.63 cycles per input byte.

jsonexamples/numbers.json
PR:
stage 1 instructions:     382783 cycles:     102428 (21.90 %) ins/cycles: 3.74 mis. branches:         74 (cycles/mis.branch 1382.45) cache accesses:       3695 (failure          0)
 stage 1 runs at 0.68 cycles per input byte.
master:
stage 1 instructions:     376916 cycles:      99805 (21.50 %) ins/cycles: 3.78 mis. branches:         85 (cycles/mis.branch 1170.68) cache accesses:       3671 (failure          0)
 stage 1 runs at 0.66 cycles per input byte.

jsonexamples/random.json
PR:
stage 1 instructions:    1988392 cycles:     549736 (44.91 %) ins/cycles: 3.62 mis. branches:        865 (cycles/mis.branch 635.27) cache accesses:      18016 (failure          9)
 stage 1 runs at 1.08 cycles per input byte.
master:
stage 1 instructions:    1988938 cycles:     549971 (44.94 %) ins/cycles: 3.62 mis. branches:        950 (cycles/mis.branch 578.32) cache accesses:      17965 (failure         14)
 stage 1 runs at 1.08 cycles per input byte.


jsonexamples/twitterescaped.json
PR:
stage 1 instructions:    1324875 cycles:     373585 (22.59 %) ins/cycles: 3.55 mis. branches:       1082 (cycles/mis.branch 345.02) cache accesses:      16359 (failure          3)
 stage 1 runs at 0.66 cycles per input byte.
master:
stage 1 instructions:    1304548 cycles:     357358 (21.52 %) ins/cycles: 3.65 mis. branches:        887 (cycles/mis.branch 402.59) cache accesses:      16361 (failure         13)
 stage 1 runs at 0.64 cycles per input byte.

jsonexamples/twitter.json
PR:
stage 1 instructions:    1722479 cycles:     504300 (51.15 %) ins/cycles: 3.42 mis. branches:        760 (cycles/mis.branch 663.15) cache accesses:      17816 (failure          1)
 stage 1 runs at 0.80 cycles per input byte.
master:
stage 1 instructions:    1706919 cycles:     477304 (49.85 %) ins/cycles: 3.58 mis. branches:        710 (cycles/mis.branch 671.84) cache accesses:      17838 (failure          0)
 stage 1 runs at 0.76 cycles per input byte.

jsonexamples/update-center.json
PR:
stage 1 instructions:    1340454 cycles:     407483 (40.52 %) ins/cycles: 3.29 mis. branches:       1597 (cycles/mis.branch 255.04) cache accesses:      16490 (failure          6)
 stage 1 runs at 0.76 cycles per input byte.
master:
stage 1 instructions:    1319994 cycles:     367585 (38.15 %) ins/cycles: 3.59 mis. branches:       1380 (cycles/mis.branch 266.26) cache accesses:      16509 (failure          2)
 stage 1 runs at 0.69 cycles per input byte.
```


In my tests, just one file is not harmed by this PR (random.json is the exception).

For ARM, this PR is apparently beneficial (except for random.json)...

```
jsonexamples/apache_builds.json
PR:
stage 1 instructions:     409500 cycles:     533827 (52.43 %) ins/cycles: 0.77 mis. branches:        444 (cycles/mis.branch 1200.17) cache accesses:      65238 (failure       2007)
 stage 1 runs at 4.19 cycles per input byte.
master:
stage 1 instructions:     421419 cycles:     544152 (51.21 %) ins/cycles: 0.77 mis. branches:        439 (cycles/mis.branch 1239.42) cache accesses:      68179 (failure       2006)
 stage 1 runs at 4.28 cycles per input byte.

jsonexamples/canada.json
PR:
stage 1 instructions:    8004293 cycles:    9736499 (39.83 %) ins/cycles: 0.82 mis. branches:      11047 (cycles/mis.branch 881.30) cache accesses:    1281225 (failure      35242)
 stage 1 runs at 4.33 cycles per input byte.
master:
stage 1 instructions:    8215316 cycles:    9911199 (40.40 %) ins/cycles: 0.83 mis. branches:      11089 (cycles/mis.branch 893.74) cache accesses:    1334134 (failure      35243)
 stage 1 runs at 4.40 cycles per input byte.

jsonexamples/citm_catalog.json
PR:
stage 1 instructions:    5293495 cycles:    6966285 (62.45 %) ins/cycles: 0.76 mis. branches:        447 (cycles/mis.branch 15570.60) cache accesses:     832181 (failure      27041)
 stage 1 runs at 4.03 cycles per input byte.
master:
stage 1 instructions:    5452917 cycles:    7111199 (63.10 %) ins/cycles: 0.77 mis. branches:        453 (cycles/mis.branch 15684.16) cache accesses:     869950 (failure      27045)
 stage 1 runs at 4.12 cycles per input byte.

jsonexamples/github_events.json
PR:
stage 1 instructions:     197619 cycles:     266184 (54.40 %) ins/cycles: 0.74 mis. branches:        263 (cycles/mis.branch 1012.06) cache accesses:      31650 (failure       1035)
 stage 1 runs at 4.09 cycles per input byte.
master:
stage 1 instructions:     203649 cycles:     271657 (53.81 %) ins/cycles: 0.75 mis. branches:        260 (cycles/mis.branch 1042.52) cache accesses:      33061 (failure       1036)
 stage 1 runs at 4.17 cycles per input byte.

jsonexamples/gsoc-2018.json
PR:
stage 1 instructions:    8335808 cycles:   12312863 (61.15 %) ins/cycles: 0.68 mis. branches:       8661 (cycles/mis.branch 1421.56) cache accesses:    1329105 (failure      52076)
 stage 1 runs at 3.70 cycles per input byte.
master:
stage 1 instructions:    8647775 cycles:   12624148 (62.24 %) ins/cycles: 0.69 mis. branches:       8779 (cycles/mis.branch 1437.96) cache accesses:    1405273 (failure      52093)
 stage 1 runs at 3.79 cycles per input byte.

jsonexamples/instruments.json
PR:
stage 1 instructions:     724617 cycles:     929828 (51.85 %) ins/cycles: 0.78 mis. branches:        833 (cycles/mis.branch 1115.42) cache accesses:     115547 (failure       3461)
 stage 1 runs at 4.22 cycles per input byte.
master:
stage 1 instructions:     745260 cycles:     947947 (53.26 %) ins/cycles: 0.79 mis. branches:        840 (cycles/mis.branch 1128.21) cache accesses:     120717 (failure       3464)
 stage 1 runs at 4.30 cycles per input byte.

jsonexamples/marine_ik.json
PR:
stage 1 instructions:   11819471 cycles:   13322990 (38.67 %) ins/cycles: 0.89 mis. branches:      13413 (cycles/mis.branch 993.22) cache accesses:    1831682 (failure      46691)
 stage 1 runs at 4.47 cycles per input byte.
master:
stage 1 instructions:   12099158 cycles:   13543186 (39.13 %) ins/cycles: 0.89 mis. branches:      13485 (cycles/mis.branch 1004.25) cache accesses:    1902389 (failure      46738)
 stage 1 runs at 4.54 cycles per input byte.

jsonexamples/mesh.json
PR:
stage 1 instructions:    2764223 cycles:    3164453 (37.36 %) ins/cycles: 0.87 mis. branches:       1327 (cycles/mis.branch 2383.53) cache accesses:     420185 (failure      11332)
 stage 1 runs at 4.37 cycles per input byte.
master:
stage 1 instructions:    2832050 cycles:    3221761 (37.99 %) ins/cycles: 0.88 mis. branches:       1363 (cycles/mis.branch 2363.54) cache accesses:     437189 (failure      11339)
 stage 1 runs at 4.45 cycles per input byte.

jsonexamples/mesh.pretty.json
PR:
stage 1 instructions:    5030116 cycles:    6415061 (51.45 %) ins/cycles: 0.78 mis. branches:        105 (cycles/mis.branch 60576.60) cache accesses:     788886 (failure      24689)
 stage 1 runs at 4.07 cycles per input byte.
master:
stage 1 instructions:    5177983 cycles:    6549022 (52.10 %) ins/cycles: 0.79 mis. branches:        164 (cycles/mis.branch 39739.21) cache accesses:     825941 (failure      24699)
 stage 1 runs at 4.15 cycles per input byte.

jsonexamples/numbers.json
PR:
stage 1 instructions:     496709 cycles:     636951 (40.37 %) ins/cycles: 0.78 mis. branches:        585 (cycles/mis.branch 1087.93) cache accesses:      79194 (failure       2362)
 stage 1 runs at 4.24 cycles per input byte.
master:
stage 1 instructions:     510764 cycles:     648139 (40.90 %) ins/cycles: 0.79 mis. branches:        573 (cycles/mis.branch 1129.56) cache accesses:      82681 (failure       2363)
 stage 1 runs at 4.32 cycles per input byte.

jsonexamples/random.json
PR:
stage 1 instructions:    2751128 cycles:    3740082 (53.05 %) ins/cycles: 0.74 mis. branches:       3163 (cycles/mis.branch 1182.30) cache accesses:     562461 (failure       8009)
 stage 1 runs at 7.33 cycles per input byte.
master:
stage 1 instructions:    2652479 cycles:    3733915 (55.58 %) ins/cycles: 0.71 mis. branches:       3162 (cycles/mis.branch 1180.74) cache accesses:     413332 (failure       8009)
 stage 1 runs at 7.31 cycles per input byte.


jsonexamples/twitterescaped.json
…